Ensure all returns have same type

+++ tests/epyccel/test_epyccel_functions.py
@@ -308,8 +308,9 @@ def test_multiple_returns_f14(language):
 def test_decorator_f15(language):
     @types('bool', 'int8', 'int16', 'int32', 'int64')
     def f15(a,b,c,d,e):
+        import numpy as np
         if a:
-            return b + c
+            return np.int64(b + c)
         else:
             return d + e
